+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> PostgreSQL components for Python Changelog
 
+## <a name="3.2.0"></a> 3.2.0 (2021-08-06)
+
+Added support for SQL schemas
+
+### Features
+* Added schema support to PostgresPersistence, IdentifiablePostgresPersistenve, IdentifiableJsonPostgresPersistence
+* Added _auto_generate_id flag to IdentifiableSqlServerPersistence
+
 ## <a name="3.1.0"></a> 3.1.0 (2021-05-14)
 
 ### Bug Fixes